Update invaders spacing
Before, the size a the square containing an invader was fixed. As a
consequence, if changing the invader design, the display needed to be
updated with other values.

Now, nothing is fixed anymore. A spacing between the invaders is defined
but the size af the invaders is also known. As a consequence, the design
of the invaders can easily be changes with no other modifications in the
code.